add controllers fn
