+++ controllers/index.js
@@ -1,5 +1,12 @@
-const contacts = require("./contacts");
-
+const getContactById = require("./contacts/getContactById");
+const listContacts = require("./contacts/listContacts");
+const addContact = require("./contacts/addContact");
+const removeContact = require("./contacts/removeContact");
+const updateById = require("./contacts/updateById");
 module.exports = {
-  contacts,
+  getContactById,
+  listContacts,
+  addContact,
+  removeContact,
+  updateById,
 };
