+++ routes/api/contacts.js
@@ -1,9 +1,9 @@
 const express = require("express");
 const router = express.Router();
-const { contacts: ctrl } = require("../../controllers/");
+const ctrl = require("../../controllers");
 const { controllerWrapper, validation } = require("../../middlewares/");
 const { contactsSchema } = require("../../schemas");
-
+console.log(ctrl);
 router.get("/", controllerWrapper(ctrl.listContacts));
 
 router.get("/:contactId", controllerWrapper(ctrl.getContactById));
